Extract {Audit,Remediate}FilePermissionsHelper functions
